feat: query builder init
