+++ src/components/DataViews/components/QueryBuilder/InitialAddButton.tsx
@@ -1,0 +1,19 @@
+import { cx } from "@emotion/css";
+import { AddIconStyle, InitialAddStyle } from "./styles";
+import AddOutlinedIcon from "@mui/icons-material/AddOutlined";
+
+export const InitialAddButton = (props: any) => {
+    const { mainTheme, resetLabelsState } = props;
+    return (
+        <div
+            className={cx(InitialAddStyle(mainTheme))}
+            onClick={resetLabelsState}
+        >
+            <AddOutlinedIcon
+                className={cx(AddIconStyle(mainTheme))}
+                fontSize="small"
+            />{" "}
+            <small>Add Labels</small>
+        </div>
+    );
+};
